squeeze unused bytes out from ftdi response
(When there are 2 bit operations in a row, this is just accumulating
shifted bits into a register for return to user.  When exiting
Shift-DR/IR state, the last bit shifted is not performed with DATAWBITS,
but with a TMS operation.  For this reason, the combo of 2 bit ops
in a row is quite common.)